requirements.txt
gizeh==0.1.11
torch==1.10.1
gym
matplotlib
seaborn

hydra-core
opencv-python
pre-commit
black
isort
pytest